fixed data processing / updated the color scales to reflect that.
I found out i was adding the total number of deaths for each day (cause i thought it was reported like that but instead i was adding the total over and and over again so it skewed the data. Fixed the color scales for each catagory to better reflect the numbers as well.